# server/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "0.0.1",
    "title": "NoCodeDB Backend",
    "description": "Backend Documentation For Your NoCodeDB API"
  },
  "tags": [
    {
      "name": "Auth",
      "description": "API For User Authentication"
    }
  ],
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "description": "Login a user and set JWT cookie for user upon login.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "description": "Email to login",
                    "example": "email@email.com"
                  },
                  "password": {
                    "type": "string",
                    "description": "Password to login",
                    "example": "password"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User has been signed in successfully."
          }
        }
      }
    },
    "/signup": {
      "post": {
        "tags": [
          "Auth"
        ],
        "description": "Signup a user, then login and set JWT cookie for user upon login.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "description": "Email to sign up",
                    "example": "email@email.com"
                  },
                  "password": {
                    "type": "string",
                    "description": "Password to sign up",
                    "example": "password"
                  },
                  "passwordConfirm": {
                    "type": "string",
                    "description": "Confirm password",
                    "example": "password"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User has been signed up successfully."
          }
        }
      }
    },
    "/token": {
      "post": {
        "tags": [
          "Auth"
        ],
        "description": "Refresh token with refresh token in cookie.",
        "responses": {
          "200": {
            "description": "Token has been refreshed successfully."
          }
        }
      }
    },
    "/logout": {
      "post": {
        "tags": [
          "Auth"
        ],
        "description": "Logout the currently signed in user.",
        "responses": {
          "200": {
            "description": "User logged out successfully."
          }
        }
      }
    }
  }
}